QToolButton a11y: Take menu from default action into account
If a menu is set, QAccessibleToolButton takes it into account
for its role, for the QAccessible::State::hasPopup state
and for reporting it as an accessible child.

So far, this was only working when the menu was
directly set using QToolButton::setMenu, but not
when the menu came from the default action set
via QToolButton::setDefaultAction.

From a user perspective, that doesn't make a
difference, so also treat these equally on
the a11y layer, by taking into account the
default action's menu as well.

Fixes: QTBUG-138172
Pick-to: 6.10 6.9
Change-Id: Ida6915e983f026db7d9840542e4b5f5c14ae13e1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>