visionOS: Avoid creating UIWindow window immersive scenes
Scenes sessions with role CPSceneSessionRoleImmersiveSpaceApplication
are backed by a CPImmersiveScene, which inherits UIWindowScene, but we
don't need our own UIWindow for these scenes, as we render the content
through compositor services.

The reason for CPImmersiveScene being a UIWindowScene is likely that it
contains a CPSceneLayerEventWindow and  UITextEffectsWindow, but these
are implementation details we shouldn't worry about.

Pick-to: 6.10 6.9 6.8
Change-Id: Ie6c492f176d6fef8853f3112f6519ca7f94952da
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>