QDBusConnection: don't disable delivery if !qApp
Amends commit 6fb39ca2cbb111469ce09c8342997eb5375e3239, which removed
the qApp check because the comment above the check didn't say it was
needed. In fact, that made the code match the comment. But if qApp is
null, we'd post an event to nullptr to reenable delivery, which meant it
was never re-enabled, resulting in stalled applications. The sequence
implied it was necessary:

    if (suspendedDelivery && result && result->connection)
        result->enableDispatchDelayed(qApp); // qApp was checked in the caller

Ideally we would make the comment true and re-enable without requiring a
QCoreApplication object, but qAddPreRoutine() takes a parameter-less
function and I'm not going to spend time adding more complexity to that.
Instead, let's just restore behavior.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-135928
Change-Id: I40c54d1343e7d0c0f5f3fffdd05e84d01d5df025
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>