QDockWidget a11y: Only report as toplevel window if floating
The QAccessible::Window role is for toplevel windows
only, and using it for non-toplevels e.g. breaks
window-relative coordinates in AT-SPI for non-floating
dock widgets and their content.

A QDockWidget is only a top-level if it is floating,
so only report the QAccessible::Window role in that
case, otherwise use QAccessible::Pane ("A generic container.")
because it's a container containing the QDockWidget::widget().

Adjust tst_QAccessibility::dockWidgetTest accordingly.

Notifying about the role change when (un)docking
a dock widget depends on introducing a new QAccessible::Event
and will be implemented separately in an upcoming commit.

Fixes: QTBUG-138206
Pick-to: 6.10 6.9 6.8
Change-Id: I9eb1bdc2a60260f169910671ebc82d9735e5ba13
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>