a11y: Introduce RoleChanged event, bridge to AT-SPI, UIA
For some accessible objects, the role isn't static,
but can change. Introduce QAccessible::RoleChanged
to allow notifying about such role changes, which allows
AT clients and platform a11y caches to be notified and
updated accordingly.

For AT-SPI, send a corresponding AT-SPI
"object:property-change:accessible-role" event.

For UIA on Windows, notify about a change of the
UIA_ControlTypePropertyId property.

The new event type will be used in an upcoming commit to
notify about role changes in QAccessibleDockWidget when
the QDockWidget gets (un)docked.

[ChangeLog][QtGui][QAccessible::Event] Added new
RoleChanged enum value that can be used to send a
corresponding event when the role of an accessible object
has changed.

Task-number: QTBUG-138206
Change-Id: Ifbd712b5573bb748be4dec9ba540f6a3a8de1c76
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>