Clarify what we mean by Julian Day in QDate and QDateTime
The formal definition of Julian Day numbers originates with
astronomers, so is actually a numbering of UTC nights: the transition
from one to the next happens at noon UTC. We, however (like many other
non-astronomical users of the numbering), do the transitions at
midnight. Make this explicit (while trying not to let it intrude too
much in the documentation).

Change-Id: I921ef5cb6507a520d0d8853f559abb07e3d2352b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>