iOS: Reflect screen changes to QWindows when view controller moves screen
On modern iOS QUIWindow and QIOSViewController is not tied to the screen
it was created on/for, as the window scene can move between the internal
screen and any external screens.

We now pick this up via the scene's update callback, as documented, and
reflect the new screen to all top level windows on the original screen.

Fixes: QTBUG-106709
Pick-to: 6.10 6.9 6.8
Change-Id: If4a2de1eec13de21fa6f2f4aecfccda345c49b16
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>