iOS: Maintain primary window scene's screen as primary screen
As long as we don't support multiple window scenes our primary
scene with role UIWindowSceneSessionRoleApplication is the one
where you can effectively place new windows, so it should be
the primary screen.

This handles the case where the primary scene is placed on an
external screen at startup, as well as when the primary scene
is moved between screens.

In the former case we also make sure to move windows on to this
screen, as they may have been created when the primary screen
was the built in display, where we don't have an associated
window scene showing.

Fixes: QTBUG-137544
Pick-to: 6.10 6.9 6.8
Change-Id: Iae87479996419ea315cece5beb88de5ccf7089ba
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>