Windows: Return hittest region for CustomizeWindowHint|ExpandedClientAreaHint
When Qt::CustomizeWindowHint in combination with Qt::ExpandedClientAreaHint is set, we are not returning any hittest region for WM_NCHITTEST, resulting in ignoring mouse events. This patch adds another code path to handle WM_NCHITTEST in this case properly. A test whether a Widget was hit is performed with QWSI with the current coordinates and button state of the mouse. In case a widget was hit the event is consumed and NCHITTEST will return HTCLIENT. In case no widget was hit, a check is performed whether the mouse event happened in the titlebar area and if so, the respective hittest region for the titlebar area is returned.

Fixes: QTBUG-137927
Pick-to: 6.10 6.9
Change-Id: I4531fec71eaaa20959a8015ea1babd6d36ee9341
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>