QTranslator: Protect QTranslator data from being accessed while altered
Protect QTranslator::translate() and the QTranslator::load() overloads
by a mutex. This in particular allows to safely call ::load() for an
object that is already installed via QCA::installTranslator().

In QTranslator::translate() we don't use a normal QMutexLocker,
but return an empty string if the mutex is locked. This is to avoid
deadlocks for code paths inside QTranslator::load() that are
'call back' via QCA::translate().

Pick-to: 6.10
Fixes: QTBUG-137179
Change-Id: Ia6d0ac58b932084886a3b32e1e14469e31acc2f0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>