Doc: Remove deadlock warnings in most QMutex documentation
The origin of this sentence predates the Qt 5 history. Anyhow, as is,
the sentence is too general, as locking the same mutex
over time from the same thread is definitely supported, as long as
it is also unlocked every time. Furthermore, for the tryLock()
variants, the thread will only dead-lock if an infinite wait time
is specified.

Improve the sentence for QMutex::lock(), but just leave out this
information for the other methods. The behavior is arguably as you
would expect it from a mutex.

Pick-to: 6.9 6.10
Change-Id: I0b658cc29c2f2d2cd6b59d49e2baa285508ec720
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>