iOS: Silence warning about stale focus object
The QPlatformInputContext contract assumes clients will call
setFocusObject before update(), but code paths in QGuiApplication
and Qt Quick will easily trigger update() for a new
qGuiApp->focusObject() before reflecting the new object via
setFocusObject().

We mitigate this on the iOS side, but we don't need a user-visible
warning for this.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-138032
Task-number: QTBUG-138256
Change-Id: I056ac3744e41811a1dab44aa50871162aef916ce
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>