Windows11Style: don't use alpha for QPalette::Base
Using a semi-transparent color for QPalette::Base is not a good idea
because this color is also used to override old content (as it's a
background color) which means the old content would shine through. To
avoid such things, workarounds were added which just created other bugs
so disable the transparency for QPalette::Base completely.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-138230
Fixes: QTBUG-135213
Task-number: QTBUG-120254
Change-Id: I6fa4d3950cd412a6b93b6b8d3634aaf685c6b959
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>