Android: simply Qt application startup logic
Currently, we have two calls responsible for application startup,
startQtAndroidPlugin() and startQtApplication().

The first one is responsible for initializing things mostly, such as
getting the main() handle, setting file engine handlers. That method is
called as a blocking task on the Qt thread.

The second one is responsible for starting the app, i.e. calling main()
and after that closing and exiting the app. This one is posted to the
Qt thread right after setting the application's state to started.

Having this separation doesn't necessarily serve much and only adds to
the complexity of the calls dance we need to do from Java <-> C++. This
patch simplifies things by having one call that does the initialization,
gets the main() handle, start the app and exit the Qt C++ application
afterwards. This will still be run under the Qt thread as before.

Change-Id: I1e93ec1a77792880426939a6302f8487a74e8915
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>