Android: simplify the Qt app's exit logic
Currently, we have terminateQt(), quitApp(), quitQt() and also
quitQtCoreApplication(). All these calls are called at some point
or another while tearing down the Qt application.

Right after main() returns, we call quitApp() which finish() the
QtActivity and sets isStarted to false, then shortly after onDestroy()
would be called, which calls terminateQt() which does the cleanup in
native C++ side and then exit()/join() the Qt thread.

The other call quitQt(), does mainly the same with the exception that
it won't finish() the Activity because it's a normal Activity and it's
owned by Qt, this is the way in Quick for Android use case.

These many calls makes the exit sequence slightly convoluted, and to
simplify it, we could do it slightly differently. Firstly, right after
main() returns, we set isStarted to false, then for normal Qt apps we
call finish() -> onDestroy() then calls terminateQt(), and for Quick for
Android apps, their delegates would call terminateQt() directly. At the
end terminateQt() would exit()/join() the Qt thread.

To follow the naming of startQtNativeApplication(), this renames
terminateQt() to terminateQtNativeApplication().

Change-Id: Iefe7fcc23039b3c58f81a8a3d7f38097f97ee07c
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>