Windows11Style: fix slider handle animation
The slider handle animation did not work correctly since it was not in
sync with the logic inside QSliderPrivate::updateHoverControl() and
determined by itself if the mouse was over the handle. If the current
value returned by subControlRect(CC_Slider, option, SC_SliderHandle) is
not suitable, then subControlRect for CC_Slider has to be reimplemented
instead.
As a drive-by clean up the code a little bit:
 - avoid a second qstyleoption_cast<const QStyleOptionSlider *>()
 - don't read _q_inner_radius from property as it's already known
 - don't enable antialiasing for QPainter - already done globally

Pick-to: 6.10 6.9
Change-Id: I742c7836736a22e02cc176e7424c4fd947e809eb
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>