QMetaObjectBuilder: Fix isRequired
There was a copy&paste error, which wasn't caught because the test
wasn't precise enough, and the actual use case in qtdeclarative doesn't
need to query the builder.

Fix that, and adjust the test to verify that setting and querying all
flags does report everything as set.

Amends 2b41b6fedeca030e0038277f78d9023d13582797

Change-Id: Iab5186d4f8aff91afc406587cc4eb11c969c4c84
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>