Disable QPlatformIntegration::Capability::OpenGL on macOS 26 in VMs
Or rather, when detecting that the software GL renderer is in use,
which is the case in VMs, or when requesting Qt::AA_UseSoftwareOpenGL
explictly.

In this situation macOS 26 crashes in [NSOpenGLContext flushBuffer]
and [NSOpenGLContext setView:], making it unusable.

Hopefully this issue is resolved in the platform, but for now we
report lack of OpenGL so we can get our tests passing/skipping in CI.

Change-Id: I68412aed5174194c80d2b079ffba141f9f2a3be1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>