tst_QTextStream: scope a variable closer (and init it)
Move the declaration of 'value' to just after the first use, in the
loop, and give it an initial value that doesn't appear in the file.

This improves test coverage by checking that the variable was indeed
overwritten by the failed op>> call, and doesn't just accidentally
have the value zero.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Task-number: QTBUG-138435
Change-Id: I4591d105dd2d991f75f0086c9706ebe9be832251
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>