tst_QTextStream: make pos3LargeFile() actually test pos()
The test function took an inordinate amount of time to execute, due to
QTBUG-138435, but didn't actually test the result of all the pos()
calls.

An earlier commit reduced the test file size by a factor of 10 to make
the runtime manageable. This patch adds actual testing of pos() by
calculating a running tally of where the expected position is, in the
file, and comparing that with QTextStream::pos().

As a drive-by, port a QString(const char*) call to a _L1 string
literal.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Task-number: QTBUG-138435
Change-Id: Ib9d4b79b68847f52a3f49ffc458098b5cabc441b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>