API Review: QRangeModel: implement roleNames as a settable property
Given use cases of QRangeModel with Qt Quick, it makes little sense to
require users to subclass QRangeModel just to get the roleNames
implementation to match the property names of gadgets used in the model.

Override roleNames and add a setter, making it a property. We can cache
the hash table at first return, avoiding the (common) mistake of
constructing a new (but generally static) has table in each call of the
getter.

Add a test and documentation.

Pick-to: 6.10
Change-Id: Ib584ae4dc227000968d49654942b71510ff4c78a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>