QFileSystemEngine/Darwin: remove superfluous check for kCFURLIsHiddenKey
As far as I can tell, CoreFramework is just testing for the UF_HIDDEN
flag being set in the st_flags result from lstat(). Since we already
perform this check ourselves, data.isHidden() will be true for every
entry for which the CF check would return true (with the previous
commit's fix applying it to lstat() instead of stat()).

This commit therefore saves us from performing an extra lstat() call and
allocating memory for the CFString and CFURL for when the entry isn't
hidden, for which CF would return false anyway.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-138374
Change-Id: I36182829ddd79374cf3cfffda6cdd44ac85f9080
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>