QFont: fix potential "Integer division by zero" in QFontCache::decreaseCache()
QFontCache::decreaseCache() may crash with "Integer division by zero"
in case when an engine is missed in engineCacheCount for some reason.
It still not clear how this may happen, but we see this in user's crash
reports.

This change simply prevents crashes, but the underlying cause of such
cache inconsistency needs to be investigated and fixed further.

Task-number: QTBUG-138455
Change-Id: I6d81c23a67a883b9c00094a8fcd5394bb86b39ea
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>