QRangeModel: add a private, move impl pointer and call logic into it
While it introduces another level of indirection for each call, it moves
implementation details out of the public header, and gives us more
options for extending the class in the future without having to change
the implementation type.

The call helpers can then move into the private, they only operate on
function pointers, and don't need to know about the specialization of
the implementation type.

Pick-to: 6.10
Change-Id: I0c6ef4ca8170929360bdaef33882acee209eb7e9
Reviewed-by: Marc Mutz <marc.mutz@qt.io>