tst_QTemporaryFile: replace custom struct with QScopeGuard
Simplify the code by using QScopeGuard. In particular, this makes
clear what "reset()" was supposed to do: dismiss.

Every Qt developer knows the semantics of QScopeGuard, but this custom
class has to be learned by every reader anew.

As a drive-by, fix the spacing inside parentheses.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I20b80fe551951a5aba01a8996c4934b574d34fac
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>