QDBusPendingReply: add tests for proper copy/move operations
Constructing a QDBusPendingReply from a QDBusPendingCall involves some
extra steps on the d_ptr. Specifically, it populates the expected reply
signature based on the provided types.

Add tests that make sure that these extra steps are actually executed,
and also check that copy/move of QDBusPedningReply does not invalidate
this reply signature.

These tests are prerequisites for the QDBusPendingReply SMF refactoring.

The tests access the internals of QDBusPendingCallPrivate, so we have
to link to QtDBusPrivate and add tst_QDBusPendingReply as a friend to
QDBusPendigCall.

Pick-to: 6.10
Change-Id: I8a2cae603321a5faa0dd1caf9401b84018617c9f
Reviewed-by: Marc Mutz <marc.mutz@qt.io>