QDBusPendingReply: apply Rule Of Zero
The only non-default SMF was the copy-assignment operator, which was
calling the assign() function.
However, that call seems to be redundant, as the other object should
have already called it during construction, and since the implementation
uses QESDP, the copy-assignment should just increase the ref count.

The tests that were added in the previous patch confirm that the call
is not needed. This means that we can apply Rule Of Zero and remove
all SMFs.

Change-Id: I436dad79730b7a8d4ef50abedc09783588a803f0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>