QString: fix overly-eager arg(int-ish) overload
In order to support (unscoped) enums, the overload accepted everything
that implicitly converted to q(u)longlong. But like the reporter of
QTBUG-138471 found out, that incorrectly matches types that implicitly
convert to float. Those types used to call the FP-ish overload set in
6.8, but that one is now restricted to actual FP types (by way of
being a template constrained, essentially, on std::is_floating_point),
and a type that converts to float is not itself a floating-point
type. Unfortunately, the int-ish overload accepted the call,
truncating the float to an int for printing.

Fix by being more explicit in the arg(int-ish) constraint, saying
exactly what we want: integrals, enums, but not scoped ones, and
nothing that already matches the string-ish or FP-ish overloads.

[ChangeLog][Potentially Source-Incompatible Changes][QtCore][QString]
Fixed the integral arg() overload to reject types that merely
implicitly convert to an integral type. This was always the intent,
but 6.9 and 6.10.0 incorrectly accepted types that implicitly convert
to float to match this overload, causing truncated results. We now
reject such types. A backwards-compatible fix is to cast such types to
a C++ type whose displayed form matches your intent.

Amends 563ed822f867c6c3040956017d4ca7f3795f172c.

Fixes: QTBUG-138471
Pick-to: 6.10
Change-Id: I6f3b6b93c243df574b878794b6304109d28d58ff
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>