wasm: make tst_qeventloop pass on webassembly
Bring the wasm event dispatcher implementation closer to
spec, as determined by tst_qeventloop. This can be done
when asyncify/jspi is enabled, if we don't have that
the behavior remains more approximate.

Make QEventDispatcherWasm::processEvents() return exactly
if it processed an event or not:
  - count the number of native events.
  - account for the internal suspend timer
  - assume that a posted event was sent if there was a
    wakeUp() call.

Send timer events in processEvents() only, if asyncify/jspi
is enabled.

Change processEvents() WaitForMoreEvents logic to:
 1) process events
 2) if there was no events then wait
 3) process events before returning

This way we only wait if there was no events (in line
with documentation), and we process events after waking
up.

Pick-to: 6.10
Change-Id: I5891b9a72d5d938b695cae9d2bfe4dcac9f62523
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>