QRangeModel: implement autoConnectPolicy
A model that operates on a range holding items of the same QObject
subclass presents the values of the properties of those QObject
instances as its data. The property names will be used as the role
names.

It is then very convenient if changes to the properties in those QObject
instances makes the model emit dataChanged() for the respective index
and role.

This requires that we identify the changed-signals for each property
that corresponds to a role name, and connect to each of those signals
for each object instance. This is an expensive operation, in particular
for trees where we have to recursively traverse the entire sturcture.
But a range holding QObjects is already quite expensive and therefore
only reasonable for small models with dozens rather than thousands of
items. At that scale, the overhead is acceptable, and we can store the
respective meta data in a baseclass of our storage type, specialized for
when the range's item type is the same QObject subclass for all columns.
We know this already, as the default implementation of roleNames() uses
that as well.

Each connection goes to a functor object that stores the index and the
role, which is the information we need to emit dataChanged. By storing
the index as a QPersistentModelIndex, objects changing position in the
model will automatically update the functor object.

The public API for this is a policy property, with two values (plus the
"None" default value): a Full auto connect iterates the entire model and
connects all relevant properties. This is expensive compared to only
connecting some objects and properties, but gives full coverage, and has
no overhead other than the connection itself.

The alternative is OnRead, which connects lazily when data is read for
the first time. This is cheaper if it's unlikely that all objects and/or
all properties are displayed, but adds book-keeping overhead so that we
know when an object and role are already connected. If everything gets
connected lazily, then this is substantially more expensive than doing
a full auto-connect in the first place.

When new rows or columns got inserted, and if autoConnectionPolicy is
set to Full, then we need to connect the new objects after QAIM emitted
rows/columnsInserted(). This gives clients a chance to populate the new
cells with objects. When rows or columns are removed and the policy is
set to OnRead, then we have to remove those connections from our book-
keeping set.

When objects are removed from the model (which means rows or columns are
removed), then we need to break the connection to the respective
objects. As the QObjects will likely be destroyed anyway, which will
then break the connections, we disconnect lazily: we disconnect if our
functor gets called for an object that has been removed. In that case,
the stored QPMI will have become invalid. To break the connection, we
need to store the QMetaObject::Connection in the functor object.

However, we only get that connection handle as a result of making the
connection, at which point our functor object has already been moved
into Qt's QCallableObject data structure. In order to store the
connection handle, we implement the move constructor of the functor to
store the address of the move-constructed functor object in the
moved-from functor object. This makes a moved-from functor act as a
reference to the moved-to functor and we can use that to store the
connection handle in the functor instance that is stored in the
QCallableObject.

Since a functor instance holds either the property data, or the address
of the moved-to instance, we can use a std::variant to avoid overhead.
Using a plain union would be even cheaper, but at this point, benefit
from the guard-rails we get from std::variant.

The context object for all connections is then a dedicated QObject,
which we can simply delete to break all connections.

[ChangeLog][Core][QRangeModel] A range model operating on a range that
holds identical QObject sublasses for all items can now automatically
connect the changed signals of all properties mapped to item roles to
the model's corresponding dataChanged signal. This allows user code to
change properties of the item-object directly, and model clients (like
item views) will get updated.

Change-Id: I742b57f0c90f705d8b7eb949ff0d026b8b4a52f3
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>