Prefer QTextStreamPrivate::write(QStringView) over (ptr, n) [2/2]: rest
Replace the remaining calls to write(const QChar *, qsizetype) with
calls that use QStringView.

Requires adapting a QChar[2] (not null-terminate, and QStringView
constructor scan for one). Chose to make it a char16_t literal
instead, and make it constexpr as a drive-by.

This removes the last caller of the write(p, n) overload, so remove
it.

Task-number: QTBUG-138520
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Iddfadc7cd7837a541e5840247f8812398807bcfe
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>