Titlecase HTTP/1 headers
Following the addition of QHttpHeaders and the relevant patches to port
existing code to use it, we decided it was OK to start sending all
header names as lower-case. This had the unfortunate side-effect that
certain HTTP/1 servers did Case-Sensitive comparisons on the header name,
causing them to stop functioning with Qt. To restore compatibility
with these, we title-case them at the boundary to HTTP/1.

Simply upper-case the first letter and the letter immediately following
a dash. Ignores specific casing of raw headers, e.g. MyHeader turns into
Myheader, and a lower-case header would still be titlecased.

Fixes: QTBUG-137203
Pick-to: 6.10 6.8
Change-Id: Ic7ccf95eb7aa5cba355fc5933b7b9c86600821bb
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>