configure: Ignore (qt_cmdline|configure).cmake from skipped repos
In a top-level build, when Qt repos are skipped with the -skip argument,
we read qt_cmdline.cmake and configure.cmake.

Now we're ignoring such files in skipped repositories.

This isn't a big problem at the moment, but it's relevant for a
subsequent change that will generate a vcpkg manifest from these files.
We don't want to add dependencies of skipped Qt repositories to a vcpkg
manifest.

Change-Id: Ib0558d972b24133ad746b6c8d18a09e11c3faf83
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>