moc: ensure printMsg() respects QByteArrayView boundaries
printMsg() receives a message as a QByteArrayView, which is not
guaranteed to be NUL-terminated. Using fprintf with %s could lead to
reading past the view's bounds.

This wasn't an issue previously, since all QByteArrayViews passed were
derived from QByteArrays or const char *, which are NUL-terminated by
design.

To future-proof the code, use %.*s with msg.size() to ensure only the
bytes inside bounds are printed.

Change-Id: I690e0dabdfc4ea9c23426ef98950845a35cd407a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>