moc: add lexem() functions that return QByteArrayViews
In many cases, the result of the lexem() functions is only used as a
view and is not stored. To avoid constructing thousands of temporary
QByteArray instances, introduce lexemView() and unquotedLexemView() that
return a QByteArrayView instead.

We need to carefuly remove .constData() calls on previously-QByteArrays,
because QByteArrayViews may not be NUL-terminated, so we must transfer
the size information to e.g. QString::fromLocal8bit().

This saves only about 3 % of temporary allocs but enables further
optimizations.

Change-Id: I55e389f7602ac42d3e2f83b38161286d5c17bb48
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>