Android: show a toast instead of a dialog in case of app start failure
Instead of showing a dialog that waits for user input to accept it to
then finish() the activity, we show a toast and right away finish().
This is a better approach for handling failed Qt libraries loading
since while the app waits for the user input, other Android callbacks
might be called (e.g. onNewIntent(), onPrepareOptionsMenu(), etc.)
and those themselves call native methods which will get an
UnsatisfiedLinkError that might hide the real failure which is that
the native libraries failed to load.

Task-number: QTBUG-132211
Task-number: QTBUG-120467
Task-number: QTBUG-70114
Task-number: QTBUG-86314
Pick-to: 6.10 6.9 6.8
Change-Id: Ib0a8a93bef7f07bbf0d24cd9860eb87e4a4fed3e
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>