QLocale: fix off-by-one error in codeToScript()
QLocale::LastScript, as the name suggests, is still a valid script,
but the old code let the loop run only while i < LastScript, so it
never checks the last entry, which therefore can't be found by string
lookup (both in codeToString() and in qt_splitLocaleName()
(ie. QLocale(QString/View))).

Fix by letting the loop run while i < NumScripts, which is a new local
constant for LastScript + 1. Also add a static_assert that the size of
script_code_list matches the cardinality of QLocale::Script, which is
actually how this bug was found...

Amends the start of the public history(!).

[ChangeLog][QtCore][QLocale] Fixed a bug where QLocale could not find
the QLocale::Script with the highest value when looked up by string
(codeToString() or QLocale(string) constructor).

Fixes: QTBUG-138566
Task-number: QTBUG-17105
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ie938588e2f6ff711a0c097803f05360104162d85
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>