QSystemLocaleData: de-pessimize constructor on Unix
The old code simply called readEnvironment() from the
constructor. That function locks the QReadWriteMutex for writing, but
the constructor call is already protected against concurrent runs by
Q_GLOBAL_STATIC, so the constructor doesn't need protection.

There's even a slight chance for deadlocks because the critical
section calls qgetenv(), which itself is mutex-protected. So if we're
called from a qEnvironmentXx() function that already holds
environmentMutex... deadlock. This can still happen on LocaleChange,
but that is much rarer than application startup.

Fix by Extract Method on readEnvironment() to create a private method
that does not lock the mutex, and then have readEnvironment() and the
constructor call it.

All these qgetenv() calls hammer on the environmentMutex good, of
course, so at some point, we should either move this code into
qenviromnentvariables.cpp (like qTzSet()) or expose environmentMutex
so the code can lock it once, do all the non-q-getenv() calls and then
unlock it once at the end. Filed QTBUG-138582 to keep track of this
issue.

Amends 5b6741c39aaf115406fedcdc15c498785db35a61.

Task-number: QTBUG-138582
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ibdd8cf4415d167aa8a88b71344e73871a0af9005
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>