QRM: remove the SingleColumn wrapper type
The RowOptions template gives us a better mechanism of declaring a type
with metaobject as a multi-role item in a list. The SingleColumn alias
of a single-element tuple is not as transparent as it should be, and
operating on a list of types wrapped this way becomes quite cumbersome
(using std::get<0>, extra braces for initialization).

Adjust the documentation to focus on the RowOptions mechanism, and
replace relevant test cases with a single-element tuple, which can still
be used for situations where specializing RowOptions is not applicable,
such as for row types that don't have a meta object.

If we see a clearer use case for a wrapper-based solution, then we can
try to add a transparent SingleColumn wrapper type for which we
specialize row_traits explicitly, without the complexity of the tuple
protocol.

Pick-to: 6.10
Change-Id: I515d9cb5c5129bd639fa9b20004be9a4101469eb
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>