Revert "Replace the manual chmod run with the file(CHMOD command"
In the situation where:
1) a project is built on macOS with a case-insensitive filesystem
2) the project has an android app name and qml module URI that differ
   only in case, e.g hangman and Hangman

What happened was that _qt_internal_android_create_runner_wrapper
tried to create a wrapper script with the name "hangman" and
then chmod it to be executable.

However because of the case-insensitive filesystem, the file never
gets created and the chmod is applied to the qml module dir Hangman,
removing write permissions, because file(CHMOD ...) overrides, and is
not additive.

This caused qmltyperegistrar to fail to write the qmltypes file into
the qml module directory during the build.

This change reverts the chmod command to the execute_process additive
chmod +x command.

It's not ideal, because it doesn't fix the underlying issue of trying
to create a wrapper with the same name / path as the qml module
directory, but at least it doesn't break the build of the project.

We don't want to rename the wrapper script to contain .sh, because
there is already Qt Creator code that expects the wrapper with the
current name.

This reverts commit 762c1561a81906ff52f85a0197fc040e21474db8.

Amends a905d26f14da1cf86b8490331c0cc41c2a45b283
Concerns f021fb74eb4a7213b573a9dfbc91320805fd3a6a

Pick-to: 6.10
Fixes: QTBUG-138565
Change-Id: Ic97081a176f6f022ed35da4e3a94385e73cdb785
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>