QLocale/ICU: static_assert(is_same<UChar,char16_t>) and remove casts
Since ICU 59, ICU's UChar is the same as char16_t in C++ mode¹, so we
don't need the casts if we use QString::DataPointer::data() instead of
QString::data() (that also avoids the detach attempt).

Add a static_assert() to document the assumption.

¹ https://icu.unicode.org/download/59 Release 59.1 on 2017-04-14,
  so should be old enough.
  (didn't find the release date of 59(.0), if there was one)

Don't fix the indentation, since there's an untouched third line in
each call and a follow-up commit will have to reindent this, anyway.

Task-number: QTBUG-138583
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I9f4a83b7b93445a32e3275ae717fb4e8f5c6a397
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>