QQueuedMetaCallEvent: don't initialize return storage if no arguments
allocArgs leaves the args_ array to be nullptr if nargs is 0. While that
should never be the case in practice, clang correctly points out the
possibility that we are dereferencing nullptr, e.g. when constructing a
QQueuedMetaCallEvent like this:

QQueuedMetaCallEvent event(0, nullptr, 0, 0, nullptr, nullptr);

Don't initialize the storage if the argument count is 0. Access to the
storage elsewhere doesn't explicitly handle index 0.

Amends f6211c079fa000c0d46b7912341f014669fa628a.

Change-Id: If42403c1fc862b8ba7456709af2456cfbec8086b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>