wasm: a11y - handle text input
The composition callbacks operate on the selected object only.
So what we need to do is to make sure that if a11y is enabled,
we do not operate on QWasmInputContext::m_inputElement.

This also means that we duplicate some code between
QWasmInputContext and QWasmAccessibility, which updates
the input element on focus changes.

Fixes: QTBUG-135590
Change-Id: I722aaccc89f300621e4d0eb74b12a93c0176c444
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>