CMake: Map paths of SBOM source files in the install prefix
When building in a non-prefix build, we put synced headers into the
qtbase build dir, aka the prefix dir. In that case these source files
are neither in the source dir nor in the build dir.

Map such paths with the /install_dir marker, to make the paths
reproducible, regardless of the build dir.

Also add a DEBUG message to more easily detect such cases in the
future if needed.

Fixes: QTBUG-136163
Pick-to: 6.8 6.9 6.10
Change-Id: Id0b515d50adc8e8da384fe481641f171517b9150
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>