QHttp2Stream: send RST_STREAM in most states when destroyed
Apart from Idle (not started) and Closed (nothing more will happen),
it is quite likely that we benefit from sending an RST_STREAM
if we are done with it in more states than just
Open (started, but no one has finished their part) or HalfClosedRemote
(the peer has finished their part).

Specifically HalfClosedLocal (we finished sending our stuff) makes a lot
of sense to send, seeing as it means we have likely sent our HEADERS
and are waiting for a response. Without marking the stream as locally
reset we would treat any incoming HEADERS as unexpected, and invalid,
closing the connection.

Fixes: QTBUG-138660
Task-number: QTBUG-138156
Pick-to: 6.10 6.9 6.8
Change-Id: I9ac6144ee85915f8f6a02e73909cdb01a73fe409
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>