windeployqt: Fix handling of special deployment of ICU libraries
ICU libraries are not Qt modules and thus should not be added to
dependentQtLibs. Adding it to dependentQtLibs will cause the "module
magic" which will result in an error because it cannot be found in our
module map.

To avoid this magic (and the error message), the ICU libraries should be
added to deployedQtLibraries directly.

Fixes: QTBUG-138474
Pick-to: 6.10
Change-Id: Ifef90066a74ae277ca4ef1775ce1b720ecbb6e40
Reviewed-by: Kai Köhne <kai.koehne@qt.io>