tst_QLocale: add a test for Greek final sigma rule
In Greek script, there are two versions of lower-case sigma: ς (GREEK
SMALL LETTER FINAL SIGMA) ("\varsigma") and σ (GREEK SMALL LETTER
SIGMA), but only one upper-case form: Σ (GREEK CAPITAL LETTER
SIGMA). The first one is used at the end of words, the middle one
everywhere else¹. So when going from upper-case to lower-case, the
algorithm needs to take into account whether the character is at the
end of the word or not.

The QLocale platform-specific implementations (ICU, Mac) get this
right, while Windows, and Qt's own implementation, which acs as
fall-back for QLocale::toUpper(), do not.

This acts as a reproducer for QTBUG-2163 and QTBUG-138705.

¹ https://en.wikipedia.org/wiki/Sigma

Task-number: QTBUG-2163
Task-number: QTBUG-138705
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I4aa44ee077de572925304c1cd957a35db19a6b2f
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>